devtools/client/shared/components/test/mochitest/test_tree_10.html

<!DOCTYPE HTML>
<html>
<!--
Test that when an item in the Tree component is expanded or collapsed the appropriate event handler fires.
-->
<head>
  <meta charset="utf-8">
  <title>Tree component test</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <link rel="stylesheet" href="chrome://devtools/skin/light-theme.css" type="text/css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript;version=1.8"></script>
<script type="application/javascript;version=1.8">
window.onload = Task.async(function* () {
  try {
    const ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
    const React = browserRequire("devtools/client/shared/vendor/react");
    const { Simulate } = React.addons.TestUtils;
    const Tree = React.createFactory(browserRequire("devtools/client/shared/components/tree"));

    const tree = ReactDOM.render(Tree(Object.assign({
      autoExpandDepth: 1
    }, TEST_TREE_INTERFACE)), window.document.body);

    yield setProps(tree, {
      focused: "A"
    });

    isRenderedTree(document.body.textContent, [
      "A:true",
      "-B:false",
      "-C:false",
      "-D:false",
      "M:false",
      "-N:false",
    ], "Should have auto-expanded one level.");
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
